Crate runtime_common

Re-exports

pub use opaque::*;

Modules

assets

authorization

benchmarks

constants

dip

KILT Decentralized Identity Provider (DIP) provider specification

errors

fees

migrations

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

pallet_id

Unique identifiers for pallets.

xcm_config

Structs

AdjustmentVariable

The adjustment variable of the runtime. Higher values will cause TargetBlockFullness to change the fees more rapidly.

BlockExecutionWeight

Time to execute an empty block. Calculated by multiplying the Average with 1.0 and adding 0.

BlockHashCount

BlockLength

Maximum length of block. Up to 5MB.

BlockWeights

Block weights base values and limits.

ExtrinsicBaseWeight

Time to execute a NO-OP extrinsic, for example System::remark. Calculated by multiplying the Average with 1.0 and adding 0.

FeeSplitRatio

Fee split ratio between treasury and block author (order is important).

MaximumMultiplier

The maximum amount of the multiplier.

MinimumMultiplier

Minimum amount of the multiplier. This value cannot be too low. A test case should ensure that combined with AdjustmentVariable, we can recover from the minimum. See multiplier_can_grow_from_zero.

RocksDbWeight

By default, Substrate uses RocksDB, so this will be the weight used throughout the runtime.

SendDustAndFeesToTreasury

TargetBlockFullness

The portion of the NORMAL_DISPATCH_RATIO that we adjust the fees with. Blocks filled less than this will decrease the weight and more will increase.

Tippers

Type Aliases

AccountId

Alias to the opaque account ID type for this chain, actually a AccountId32. This is always 32 bytes.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

AccountPublic

Alias to the public key used for this chain, actually a MultiSigner. Like the signature, this also isn’t a fixed size when encoded, as different cryptos have different size public keys.

Amount

AuthorityId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

BlockNumber

An index to a block.

ChainId

Identifier for a chain. 32-bit should be plenty.

DidIdentifier

A Kilt DID subject identifier.

DigestItem

Digest item type.

FeeSplit

Split the fees using a preconfigured Ratio (runtime_common::FeeSplitRatio).

Hash

A hash of some data used by the chain.

Hasher

Hasher for chain data.

NegativeImbalanceOf

Nonce

Nonce of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

SlowAdjustingFeeUpdate

Parameterized slow adjusting fee updated based on https://w3f-research.readthedocs.io/en/latest/polkadot/Token%20Economics.html#-2.-slow-adjusting-mechanism