Module runtime_common::apis

Runtime apis useful in the Centrifuge ecosystem

Enums

RewardDomain

Traits

AccountConversionApi

Runtime Api for the pallet-anchors, to be implemented by and for a specific runtime that uses that pallet.

AnchorApi

Runtime Api for the pallet-anchors, to be implemented by and for a specific runtime that uses that pallet.

InvestmentsApi

Runtime API for investments

LoansApi

Runtime API for the rewards pallet.

PoolsApi

Runtime for pallet-pool-system.

RewardsApi

Runtime API for the rewards pallet.