Module runtime_common::constants

Modules

attestation

delegation

deposit_storage

did

did_lookup

dip_provider

fee

governance

multisig

pallet_migration

preimage

proxy

public_credentials

staking

tips

treasury

web3_names

Structs

ByteDeposit

Deposits per byte

CtypeFee

CType Pallet. Per byte fee for a ctype.

IndicesDeposit

Index Pallet. Deposit taken for an account index

MinVestedTransfer

Vesting Pallet. Copied from Kusama & Polkadot runtime

UnvestedFundsAllowedWithdrawReasons

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~10% of the block weight is consumed by on_initalize handlers. This is used to limit the maximal weight of a single extrinsic.

BLOCKS_PER_YEAR

BLOCK_PROCESSING_VELOCITY

How many parachain blocks are processed by the relay chain per parent. Limits the number of blocks authored per slot.

DAYS

DEPOSIT_STORAGE_BYTE

Deposit that must be provided for each occupied storage byte.

DEPOSIT_STORAGE_ITEM

Deposit that must be provided for each occupied storage item.

EXISTENTIAL_DEPOSIT

HOURS

INFLATION_CONFIG

KILT

One KILT

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 12 second average block time.

MAX_COLLATOR_STAKE

MAX_INDICES_BYTE_LENGTH

The size of an index in the index pallet. The size is checked in the runtime by a test.

MAX_VESTING_SCHEDULES

Copied from Kusama & Polkadot runtime

MICRO_KILT

0.000_001 KILT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targetting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLI_KILT

0.001 KILT

MINUTES

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

PRIMARY_PROBABILITY

RELAY_CHAIN_SLOT_DURATION_MILLIS

Relay chain slot duration, in milliseconds.

SLOT_DURATION

UNINCLUDED_SEGMENT_CAPACITY

Maximum number of blocks simultaneously accepted by the Runtime, not yet included into the relay chain.

Functions

deposit

Calculate the storage deposit based on the number of storage items and the combined byte size of those items.

kilt_inflation_config

Inflation configuration which is used at genesis