Module runtime_common::evm::precompile

Structs

CentrifugePrecompiles

Constants

BLAKE2F_ADDR

BN128ADD_ADDR

BN128MUL_ADDR

BN128PAIRING_ADDR

DISPATCH_ADDR

ECRECOVERPUBLICKEY_ADDR

ECRECOVER_ADDR

IDENTITY_ADDR

LP_AXELAR_GATEWAY

The address of our local Axelar gateway. This is the address that Liquidity-Pool contracts on other domains must use in order to hit the Liquidity-Pool logic on centrifuge.

MODEXP_ADDR

PRECOMPILE_CODE_STORAGE

pallet_evm::AccountCodes must be populated for precompiles as otherwise OPCODE::EXTCODESIZE will make the EVM error upon calling an precompile.

RIPEMD160_ADDR

SHA3FIPS256_ADDR

SHA256_ADDR

Type Aliases

Altair

Altair’s precompiles For now, Altair uses the exact same set of precompiles used in Development.

Development

A set of precompiles. This set might contain not yet mainnet ready precompiles in order to test those in development or staging environment without touching the mainnet set.