Type Alias runtime_common::liquidity_pools::LiquidityPoolsMessage

pub type LiquidityPoolsMessage = Message<Domain, PoolId, TrancheId, Balance, Ratio>;

Aliased Type

enum LiquidityPoolsMessage {

    Invalid,
    AddCurrency {
        currency: u128,
        evm_address: [u8; 20],
    },
    AddPool {
        pool_id: u64,
    },
    AllowInvestmentCurrency {
        pool_id: u64,
        currency: u128,
    },
    AddTranche {
        pool_id: u64,
        tranche_id: [u8; 16],
        token_name: [u8; 128],
        token_symbol: [u8; 32],
        decimals: u8,
        restriction_set: u8,
    },
    UpdateTrancheTokenPrice {
        pool_id: u64,
        tranche_id: [u8; 16],
        currency: u128,
        price: FixedU128<1000000000000000000>,
        computed_at: u64,
    },
    UpdateMember {
        pool_id: u64,
        tranche_id: [u8; 16],
        member: [u8; 32],
        valid_until: u64,
    },
    Transfer {
        currency: u128,
        sender: [u8; 32],
        receiver: [u8; 32],
        amount: u128,
    },
    TransferTrancheTokens {
        pool_id: u64,
        tranche_id: [u8; 16],
        sender: [u8; 32],
        domain: Domain,
        receiver: [u8; 32],
        amount: u128,
    },
    IncreaseInvestOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        amount: u128,
    },
    DecreaseInvestOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        amount: u128,
    },
    IncreaseRedeemOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        amount: u128,
    },
    DecreaseRedeemOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        amount: u128,
    },
    CollectInvest {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
    },
    CollectRedeem {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
    },
    ExecutedDecreaseInvestOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        currency_payout: u128,
        remaining_invest_amount: u128,
    },
    ExecutedDecreaseRedeemOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        tranche_tokens_payout: u128,
        remaining_redeem_amount: u128,
    },
    ExecutedCollectInvest {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        currency_payout: u128,
        tranche_tokens_payout: u128,
        remaining_invest_amount: u128,
    },
    ExecutedCollectRedeem {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
        currency_payout: u128,
        tranche_tokens_payout: u128,
        remaining_redeem_amount: u128,
    },
    CancelInvestOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
    },
    CancelRedeemOrder {
        pool_id: u64,
        tranche_id: [u8; 16],
        investor: [u8; 32],
        currency: u128,
    },
    ScheduleUpgrade {
        contract: [u8; 20],
    },
    CancelUpgrade {
        contract: [u8; 20],
    },
    UpdateTrancheTokenMetadata {
        pool_id: u64,
        tranche_id: [u8; 16],
        token_name: [u8; 128],
        token_symbol: [u8; 32],
    },

}

Variants

Invalid

AddCurrency

Fields

currency: u128

evm_address: [u8; 20]

Add a currency to a domain, i.e, register the mapping of a currency id to the corresponding EVM Address.

Directionality: Centrifuge -> EVM Domain.

AddPool

Fields

pool_id: u64

Add a pool to a domain.

Directionality: Centrifuge -> EVM Domain.

AllowInvestmentCurrency

Fields

pool_id: u64

currency: u128

Allow a currency to be used as a pool currency and to invest in a pool.

Directionality: Centrifuge -> EVM Domain.

AddTranche

Fields

pool_id: u64

tranche_id: [u8; 16]

token_name: [u8; 128]

token_symbol: [u8; 32]

decimals: u8

restriction_set: u8

The RestrictionManager implementation to be used for this tranche token on the domain it will be added and subsequently deployed in.

Add a tranche to an already existing pool on the target domain. The decimals of a tranche MUST be equal to the decimals of a pool. Thus, consuming domains MUST take care of storing the decimals upon receiving an AddPool message.

Directionality: Centrifuge -> EVM Domain.

UpdateTrancheTokenPrice

Fields

pool_id: u64

tranche_id: [u8; 16]

currency: u128

price: FixedU128<1000000000000000000>

computed_at: u64

The timestamp at which the price was computed

Update the price of a tranche token on the target domain.

Directionality: Centrifuge -> EVM Domain.

UpdateMember

Fields

pool_id: u64

tranche_id: [u8; 16]

member: [u8; 32]

valid_until: u64

Whitelist an address for the specified pair of pool and tranche token on the target domain.

Directionality: Centrifuge -> EVM Domain.

Transfer

Fields

currency: u128

sender: [u8; 32]

receiver: [u8; 32]

amount: u128

Transfer non-tranche tokens fungibles. For v2, it will only support stable-coins.

Directionality: Centrifuge <-> EVM Domain.

NOTE: Receiving domain must not accept every incoming token. For Centrifuge -> EVM Domain: AddCurrency should have been called beforehand. For Centrifuge <- EVM Domain: We can assume AddCurrency has been called for that domain already.

TransferTrancheTokens

Fields

pool_id: u64

tranche_id: [u8; 16]

sender: [u8; 32]

domain: Domain

receiver: [u8; 32]

amount: u128

Transfer tranche tokens between domains.

Directionality: Centrifuge <-> EVM Domain.

IncreaseInvestOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

amount: u128

Increase the invest order amount for the specified pair of pool and tranche token.

Directionality: Centrifuge <- EVM Domain.

DecreaseInvestOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

amount: u128

Reduce the invest order amount for the specified pair of pool and tranche token.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

IncreaseRedeemOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

amount: u128

Increase the redeem order amount for the specified pair of pool and tranche token.

Directionality: Centrifuge <- EVM Domain.

DecreaseRedeemOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

amount: u128

Reduce the redeem order amount for the specified pair of pool and tranche token.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

CollectInvest

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

Collect the investment for the specified pair of pool and tranche token.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

CollectRedeem

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

Collect the proceeds for the specified pair of pool and tranche token.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

ExecutedDecreaseInvestOrder

Fields

pool_id: u64

The pool id

tranche_id: [u8; 16]

The tranche id

investor: [u8; 32]

The investor’s address

currency: u128

The currency in which DecreaseInvestOrder was realised

currency_payout: u128

The amount of currency that was actually executed in the original DecreaseInvestOrder message, i.e., the amount by which the investment order was actually decreased by.

remaining_invest_amount: u128

The remaining investment amount denominated in the foreign payment currency. It reflects the sum of the unprocessed as well as the processed but not yet collected amounts.

The message sent back to the domain from which a DecreaseInvestOrder message was received, ensuring the correct state update on said domain and that the investor’s wallet is updated accordingly.

Directionality: Centrifuge -> EVM Domain.

ExecutedDecreaseRedeemOrder

Fields

pool_id: u64

The pool id

tranche_id: [u8; 16]

The tranche id

investor: [u8; 32]

The investor’s address

currency: u128

The currency in which DecreaseRedeemOrder was realised

tranche_tokens_payout: u128

The amount of tranche tokens that was actually executed in the original DecreaseRedeemOrder message, i.e., the amount by which the redeem order was actually decreased by.

remaining_redeem_amount: u128

The remaining redemption amount. It reflects the sum of the unprocessed as well as the processed but not yet collected amount of tranche tokens.

The message sent back to the domain from which a DecreaseRedeemOrder message was received, ensuring the correct state update on said domain and that the investor’s wallet is updated accordingly.

Directionality: Centrifuge -> EVM Domain.

ExecutedCollectInvest

Fields

pool_id: u64

The pool id

tranche_id: [u8; 16]

The tranche

investor: [u8; 32]

The investor’s address

currency: u128

The currency in which the investment was realised

currency_payout: u128

The amount that was actually collected, in currency units

tranche_tokens_payout: u128

The amount of tranche tokens received for the investment made

remaining_invest_amount: u128

The remaining investment amount denominated in the foreign payment currency. It reflects the sum of the unprocessed as well as the processed but not yet collected amounts.

The message sent back to the domain from which a CollectInvest message has been received, which will ensure the investor gets the payout respective to their investment.

Directionality: Centrifuge -> EVM Domain.

ExecutedCollectRedeem

Fields

pool_id: u64

The pool id

tranche_id: [u8; 16]

The tranche id

investor: [u8; 32]

The investor’s address

currency: u128

The stable coin currency in which the payout takes place

currency_payout: u128

The amount of currency being paid out to the investor

tranche_tokens_payout: u128

How many tranche tokens were actually redeemed

remaining_redeem_amount: u128

The remaining redemption amount. It reflects the sum of the unprocessed as well as the processed but not yet collected amount of tranche tokens.

The message sent back to the domain from which a CollectRedeem message has been received, which will ensure the investor gets the payout respective to their redemption.

Directionality: Centrifuge -> EVM Domain.

CancelInvestOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

Cancel an unprocessed invest order for the specified pair of pool and tranche token.

Special instance of DecreaseInvestOrder where the amount is chosen properly to cancel out the ongoing investment. Required for ERC4646.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

CancelRedeemOrder

Fields

pool_id: u64

tranche_id: [u8; 16]

investor: [u8; 32]

currency: u128

Reduce the redeem order amount for the specified pair of pool and tranche token.

Special instance of DecreaseRedeemOrder where the amount is chosen properly to cancel out the ongoing redemption. Required for ERC4646.

On success, triggers a message sent back to the sending domain. The message will take care of re-funding the investor with the given amount the order was reduced with. The investor address is used as the receiver of that tokens.

Directionality: Centrifuge <- EVM Domain.

ScheduleUpgrade

Fields

contract: [u8; 20]

The EVM contract address

Schedules an EVM address to become rely-able by the gateway. Intended to be used via governance to execute EVM spells.

Directionality: Centrifuge -> EVM Domain.

CancelUpgrade

Fields

contract: [u8; 20]

The EVM contract address

Cancel the scheduled process for an EVM address to become rely-able by the gateway. Intended to be used via governance to execute EVM spells.

Directionality: Centrifuge -> EVM Domain.

UpdateTrancheTokenMetadata

Fields

pool_id: u64

tranche_id: [u8; 16]

token_name: [u8; 128]

token_symbol: [u8; 32]

Updates the name and symbol of a tranche token.

NOTE: We do not allow updating the decimals as this would require migrating all associated balances.

Directionality: Centrifuge -> EVM Domain.