Module runtime_common::migrations

Migrations

Structs

AssetManagerChangeStateminePrefixes

AssetManagerPopulateAssetTypeIdStorage

AssetManagerUnitsWithAssetType

AuthorMappingAddKeysToRegistrationInfo

This module acts as a registry where each migration is defined. Each migration should implement the “Migration” trait declared in the pallet-migrations crate. A moonbeam migration wrapping the similarly named migration in pallet-author-mapping

AuthorSlotFilterEligibleRatioToEligiblityCount

BaseFeePerGas

BaseFee pallet, set missing storage values.

CommonMigrations

MigrateBaseFeePerGas

MigrateCollectivePallets

ParachainStakingIncreaseMaxDelegationsPerCandidate

Staking increase max counted delegations per collator candidate

ParachainStakingPatchIncorrectDelegationSums

Patch delegations total mismatch

ParachainStakingPurgeStaleStorage

A moonbeam migration wrapping the similarly named migration in parachain-staking

ParachainStakingSplitCandidateState

Staking split candidate state

ParachainStakingSplitDelegatorStateIntoDelegationScheduledRequests

Staking split delegator state into [parachain_staking::DelegatorScheduledRequests]

SchedulerMigrationV3

XcmMigrations

XcmPaymentSupportedAssets

XcmTransactorMaxTransactWeight