Crate runtime_common

Common types and primitives used for Centrifuge chain runtime.

Re-exports

pub use apis::*;

pub use constants::*;

pub use types::*;

Modules

apis

Runtime apis useful in the Centrifuge ecosystem

asset_registry

AssetRegistry’s AssetProcessor

constants

Common constants for all runtimes

parachains

Listing of parachains we integrate with. For each parachain, we are interested in stating their parachain ID as well as any of their token key ID that we possibly support in our XCM configuration. These token keys are defined in said parachain and must always match the value there defined, which is expected to never change once defined since they help define the canonical id of said tokens in the network, which is relevant for XCM transfers.

types

Common types for all runtimes

xcm

xcm_fees

Structs

AssetInfo

All data for an instance of an NFT.

DealWithFees

TrancheToken

WeightToFee

Handles converting a weight scalar to a fee value, based on the scale and granularity of the node’s balance type.

Enums

CurrencyId

Type Definitions

NegativeImbalance