Module runtime_common::constants

Re-exports

pub use currency::*;

Modules

currency

Structs

AssetUnbondingSlashingSpans

BaseWithdrawalFee

BaseXcmWeight

BasicPerSecond

BlockHashCount

Burn

CanEncodeAsset

Days

DisabledValidatorsThreshold

DustRemovalWhitelist

ExistentialDeposit

Same as Polkadot Relay Chain.

ExistentialDeposits

FeedLimit

FeedPalletId

IndexTokenLockIdentifier

LockupPeriod

MaxActiveDeposits

MaxApprovals

MaxCandidates

MaxDecimals

MaxInstructions

MaxInvulnerables

MaxLocks

MinCandidates

MinCouncilMembers

MinCouncilVotes

MinimumBondExtra

MinimumPeriod

MinimumRedemption

MinimumRemoteReserveBalance

MinimumReserve

MinimumStatemintTransferAmount

Offset

OperationalFeeMultiplier

OracleLimit

PINTAssetId

PalletIndexStringLimit

Period

PintTreasuryAccount

PotId

ProposalBond

ProposalBondMinimum

RedemptionFee

RelayChainAssetId

RelayChainOrigin

RelayLocation

ReservedDmpWeight

ReservedXcmpWeight

RuntimeBlockLength

RuntimeBlockWeights

SS58Prefix

SpendPeriod

StringLimit

TransactionByteFee

TreasuryPalletId

UncleGenerations

UnitWeightCost

WithdrawalPeriod

Constants

AVERAGE_ON_INITIALIZE_RATIO

We assume that ~10% of the block weight is consumed by on_initalize handlers. This is used to limit the maximal weight of a single extrinsic.

DAYS

HOURS

KUSAMA_BONDING_DURATION_IN_BLOCKS

KUSAMA_EPOCH_DURATION_IN_SLOTS

MAXIMUM_BLOCK_WEIGHT

We allow for 0.5 seconds of compute with a 6 second average block time.

MICROUNIT

MILLISECS_PER_BLOCK

This determines the average expected block time that we are targeting. Blocks will be produced at a minimum duration defined by SLOT_DURATION. SLOT_DURATION is picked up by pallet_timestamp which is in turn picked up by pallet_aura to implement fn slot_duration().

MILLIUNIT

MINUTES

NORMAL_DISPATCH_RATIO

We allow Normal extrinsics to fill up the block up to 75%, the rest can be used by Operational extrinsics.

POLKADOT_BONDING_DURATION_IN_BLOCKS

28 eras for unbonding (28 days). Six sessions in an era (24 hours)

POLKADOT_EPOCH_DURATION_IN_SLOTS

PRIMARY_PROBABILITY

SLOT_DURATION

UNIT

WEEKS

Functions

get_all_pallet_accounts